refactor: 경고 수정 및 코드 개선
- LottoResult에서 묵시적 형 변환을 명시적으로 캐스팅
- OutputView에서 중복된 String.format() 호출을 제거

+++ src/main/java/lotto/domain/LottoResult.java
@@ -32,9 +32,10 @@ public class LottoResult {
     private long calculateTotalPrize() {
         long total = 0;
         for (Rank rank : Rank.values()) {
-            total += rank.getPrize() * rankCounts.get(rank);
+            total += (long) rank.getPrize() * rankCounts.get(rank);
         }
         return total;
     }
+
 }
